Add metrics(euclidean and cosin distance)

+++ src/main/java/ru/mtl/VoidVoice/model/Point3d.java
@@ -15,13 +15,13 @@ public class Point3d implements PersistentEntity<Long>, Comparable<Point3d> {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private long id;
 
-    private float x;
-    private float y;
-    private float z;
+    private double x;
+    private double y;
+    private double z;
 
     public Point3d() {}
 
-    public Point3d(float x, float y, float z){
+    public Point3d(double x, double y, double z){
         this.x = x;
         this.y = y;
         this.z = z;
@@ -39,27 +39,27 @@ public class Point3d implements PersistentEntity<Long>, Comparable<Point3d> {
     }
 
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public float getZ() {
+    public double getZ() {
         return z;
     }
 
-    public void setZ(float z) {
+    public void setZ(double z) {
         this.z = z;
     }
